feat: Add TransactionPage component for managing user transactions
- Implemented TransactionPage with transaction fetching, filtering, and pagination.
- Added search functionality for order ID, meal, and seller.
- Enhanced UI with status badges and detailed transaction information.
- Updated SidebarNavbar for improved layout and responsiveness.
- Fixed minor issues in AnalyticsDashboard with useEffect dependencies.
